Fix outgoing jabber logging.
Added two new perlglue functions:

* log_message - takes a message hash, turns it into an owl message,
and passes it to the logger.

* add_and_log_message - combination off add_message and
log_message. Takes a message hash, turns it into an owl message, logs
it, and adds the same message to the message list if needed. This
exists for convenience, so we don't have to convert the message hash
twice.

Also, took out an extraneous check from logging.c.