Fix a bug rendering multi-byte characters in the last column of the window.
When the last character on a line was not wide, the code assumed that
it consumed one byte. This caused multi-byte characters to render
improperly. Now we find the start of the next character to ensure we
get the whole thing. [Fixes: #56]

Reported by: andersk

Also, clarified some comments and did a bit of code hoisting.