Don't fill-paragraph the ending dot of the buffer.
When I leave a dot at the end of the message and then go back and edit
it a bit before sending, it's annoying for M-q to wrap the dot onto
the paragraph.  It really isn't part of the paragraph, so with this
patch we don't treat it as one.

Signed-off-by: Greg Price <price@mit.edu>