Disable cdbs configure handling
CDBS fails to handle aclocal correctly because it will only create it
if it exists.  The upstream tarball has a configure.  WE'll use that
and run autoreconf (and include the results in the diff) if necessary.

Also, disable dh_clean removal of autom4te.cache.  It's broken that
the upstream tarball includes this, but cleaning it up produces
problems for the diff and git-buildpackage.