Preliminary support for other Twitter-compatible microblogging services, eg. identi.ca.
Adds an optional 'service' parameter to the configuration file, which points to the API
root URL (defaults to Twitter's).  Also adds optional 'apihost' and 'apirealm' parameters
for the domain name and port of the site to connect to, and the name of the API realm,
respectively (default to Twitters'; if service is not Twitter, attempt to guess sane
defaults).