Include non-default class in personal_context
I think this turns out to be a no-op, because
only the default class is ever set. Regardless,
this feels like the right thing to do.

Also, slightly refactors to make getting a
-c class -i instance string without default class
or instance easier.

Further, quotes instance in personal context,
which seems reasonable since the syntax seems
intended to be that of zwrite.