Allow SIGINT to interrupt getting the Zephyr buddy list
If we're behind a NAT or Zephyr is losing for some other reason, 'l' will
hang for (30 seconds) * (size of .anyone). Allow SIGINT to interrupt it
and cause us to bail out early. ZLocateUser will still hang for up to 30
seconds before that takes effect, but that's better than having to wait
for all of them.