editwin: Properly update points inside a replaced region.
If the point is inside a region that is replaced, the correct thing to
do is to send it to the new end, not the old end.

Moreover, if the region shrunk and was near the end of the buffer, the
old behavior would even result in e->index > e->bufflen, which could
result in a crash.