Don't suggest repeated flags in complete_flags
For most commands, it doesn't make sense to provide an option twice, so
we avoid offering it again as a completion. Note: if the user happened
to pass the flag again, we'll still attempt to complete it.

To disable, use the repeat_flags keyword option. Tests update, with
complete_zwrite testing non-repeating and complete_word testing the old
behavior.

Signed-off-by: David Benjamin <davidben@mit.edu>