Return unicode strings to perl.
Define a new typedef, 'utf8', in perlglue.xs, and use a typemap that
indicates that values of type utf8* should be returned to perl as
Unicode strings.

There is no good reason that get_{data,config}_dir should not also
return unicode, except that making them do so crashes my barnowl in
some extremely strange and hard-to-debug manner.