Add a pre-select action list.
Allow us to add actions that should be performed before calling
pselect(). If the action performs any useful work, it should return
non-zero. If any of the actions return a non-zero value, the timeout
for pselect() will be set to 0s, so that we can process all of these
actions again in case they are interacting with one another, while
still keeping an eye on our file descriptors and timers.