Prefer popwin to typwin for cursor placement
If we somehow have a popwin open while the typwin is active, the popwin
should have preferred cursor position. This makes the cursor position
match the ordering in the old owl_process_input_char.

We should probably associate a WINDOW with each context, but that might
want to wait until we get a wrapper that persists across resizes. PANEL
might work, but if we ever use subwins...

Signed-off-by: David Benjamin <davidben@mit.edu>