Send multiple PRIVMSGs for IRC messages entered as multiple paragraphs
It is occasionally useful to be able to send multiple lines of IRC text
at once, and I've grown used to barnowl's excellent support for
composing long messages. As a compromise between allowing multiline IRC
messages and not causing autowrap to send them inadvertently, send one
IRC-level message (PRIVMSG or CTCP ACTION) for each
two-newlines-separated paragraph in the user input to :irc-msg.

Reviewed-by: Geoffrey Thomas <geofft@mit.edu>