Mark variables modified from signal handlers as volatile sig_atomic_t.
According to POSIX, this is the only type of variable (other than
errno) that may be accessed from a signal handler with defined
results.  In practice, at least the volatile is important to prevent
the compiler from reordering accesses.

err_signal_info is a bigger struct that cannot be assigned atomically,
but it’s only for debugging so we just do the best we can.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>
Reviewed-by: Nelson Elhage <nelhage@mit.edu>