Require resizeterm
This allows us to get rid of the endwin, which causes problems.
Specifically that ncurses will query the size itself on the next
doupdate, causing a potential race. It also results in an annoying
flicker that we see in typwindelta.

If we punt passing arguments to owl_global_resize, I think we can
support resizeterm-less curses with a bit of a dance:

  endwin()
  h,w = query_terminal_size()
  relayout_windows(h, w)
  doupdate()
  if (h,w != query_ncurses_size())
      h,w = query_ncurses_size()
      relayout_windows(h, w)

(This will only flicker given a race.) That said, this is annoying and
we build with our own curses on athena 9. We may as well just assume
it's there.