Actually initialize owl_global in tester.c
Fixes tests relying on variables being initialized. Better not to keep
duplicating all the owl_global code. We tell ncurses to open a fake
screen a /dev/null and run with it. Explicitly pass "xterm" as the
terminal type to reduce reliance on the running environment. (Passing
NULL defaults to getenv("TERM").)

Signed-off-by: David Benjamin <davidben@mit.edu>