A fairly hacky job of porting sepbar to owl_window
We now use the new framework, but the sepbar isn't being marked dirty at
the appropriate times. It'll want to be redone later to take advantage
of the new signals.

The 'in' parameter was also dropped. It was being unused. At best it
provided a strange bug when resizing with appendtosepbar set.

Also revert the attempt at cleaning up the cursor. We still have a hack,
but using owl_window. The cursor probably needs integration with the
context stack or some such.