Punted __unspecified__ default value for default_style.
variable.c set default_style to __unspecified__, and owl.c checked
this and then changed it to default.  This seemed to have no purpose,
and additionally would cause problems if the user decided to make a
custom style and named it __unspecified__, so I had variable.c set
default_style to default and removed the code from owl.c.