Fixed :unsub to make it work if ~/.zephyr.subs is a symlink (trac 133)
link() fails on AFS if the target is a symlink.  Additionally, on
systems where link() succeeds, barnowl overwrites the symlink, instead
of the actual file.  To fix this, I've changed owl_util_file_deleteline
to resolve symlinks on the passed file name first.  The new (backup)
file is written in the resolved directory, not the directory where the
symlink itself is found.

A helper function, owl_util_recursive_resolve_link, is added to
recursivly resolve symlinks (if necessary).

Reviewed-By: Karl Ramm <kcr@1ts.org>