Fix up feature macros somewhat
Set _XOPEN_SOURCE to 500 to target a sufficiently new standard. Also
define _BSD_SOURCE because zephyr needs a BSD extension (caddr_t).
Apparently both of these bugs were masked on Debathena because Ubuntu's
perl defines _GNU_SOURCE which turns everything on.

This also removes the need for __EXTENSIONS__ to get struct timeval on
Solaris. But we can't quite get rid of it yet because we use strcasecmp
and MAXHOSTNAMELEN.

Reported-By: Alex Vandiver <alexmv@mit.edu>