Append sender's realm to Cc'd unqualified names when replying
Currently, if you reply to a personal from a user in another realm,
that message was Cc'd to other people on their realm, and the sender
didn't specify their realm, Barnowl will try to Cc the message to
users of the same name in your realm. So, if you are 'bar@EXAMPLE.ORG'
and you get a message from 'foo@EXAMPLE.COM' cc'd to 'bletch', if you
try to reply, Barnowl will address it to foo@EXAMPLE.COM, cc'd to
'bletch', which would be bletch@EXAMPLE.ORG, which is clearly
wrong. This patch appends the sender's realm to unqualified names in
the message's Cc line when generating the reply zwrite command line.

[davidben@mit.edu: Fix up some edge case handling]