Only handle CCs in messages sent directly to you.
A zephyr with a CC line received on @SOME-REALM would still handle the
CC logic in replycmd. This oddity could possibly be retrofited to allow
@SOME-REALM to participate in CC'd messages, but that would require us
to handle CC lines on messages received on *. In the meantime, let's not
leave a half-baked not-feature.

Also refactor the block slightly. The original formulation is confusing;
it means to condition on the recipient of the message we reply to, but
actually checks $to, the recipient we use in the reply. It just happens
that in one case, the two are equal. The original code would assume that
case and then change its mind if it was wrong.