Add a zephyr GSource
This may actually have some use as a GSource instead of g_idle_add since
we queue messages up and only process so many at a time. The use of
ZQLength isn't strictly necessary, but we may as well be fancy and wait
for the poll() to tell us if there's any point in reading before doing
so.

This is the third pre-select action.