Cleanly destroy the signal thread on shutdown
This isn't strictly necessary here, but we should probably be in the
habit of safely cleaning up our helper threads.

Switch the implementation from GThread to pthreads directly. The
cleanest way to kill a signaling thread is to send it SIGTERM with
pthread_kill, but GThread doesn't expose that and gives no way to get at
the pthread_t.