Punt a number of g_new NULL checks, various minor cleanups
No real need to check g_new's return value as it always succeeds. (Or
destroys the universe^Wprocess in embarrassment.) Also make owl_global
manage the lifetime of the kill buffer since it's part of owl_global
now. And replace a g_new/strncpy with a g_strndup. It's shorter.