Show how far you are in a long message
Addresses most of ticket #119 - it shows how far you are in a long
message, but not in a message that you're typing.  The display style
I've somewhat arbitrarily chosen is to replace <truncated> with
<truncated: $num_lines_through/$total_lines>.

Additionally, the <truncated> part of sepbar has been moved to be next
to 'SCROLL', which it is most similar to.