Ability to set exposure arbitrarily (like "zctl set exposure")
Zephyr variables 'exposure' and 'default_exposure' have been added.  The
'exposure' variable defaults to the value of 'default_exposure', which
defaults to the value in ~/.zephyr.vars, if there is one, or to
realm-visible.  When the value of default_exposure is set with :set
default_exposure ..., this value is written to ~/.zephyr.vars.

Additionally, OWLVAR_STRING_FULL has been modified to allow custom
setting of validset.

This fixes ticket # 65.