Use g_strjoinv in owl_cmddict_execute_argv, don't quote correctly
This reverts commit e3c8332fa85642544dba1222912b77cf6e32ce8c and uses
g_strjoinv instead. The correctly quoted form caused the j keybinding in
BarnOwl::Jabber to break. Without it, the world isn't sane this way
either:

  :perl BarnOwl::bindkey(recv => 'C-f C-f' => command => 'recv:next')

But a solution to this will probably be far more complex and bear little
relation to this commit. In the meantime master really shouldn't have
this obnoxious of a regression on it.