Properly discover things when /usr/athena != /usr
On Solaris, by an accident of configure and krb5-config, we manage to
discover libzephyr and libcom_err, but not com_err.h. Thankfully the
latter is not fatal and we relied on C's defaulting everything to int.
And by accident we were searching for libzephyr, and not zephyr.h.

Also use CPPFLAGS instead of CFLAGS to pull in our bundled header files,
otherwise autoconf warns about a discrepancy between the preprocessor
and compiler.