Just require stdbool.h and tell ncurses to use it
Now that Solaris' stdbool.h works for us, we may as well just require
it. The preamble is repetitive and didn't actually work anyway. (It
conflicts with ncurses' attempts to redefine bool when stdbool.h is
unavailable or disabled.)

We may need to add a different preamble later if we care about including
these headers in C++ (Solaris has a C++-incompatible stdbool.h), but we
can deal with that later.