Fixed a bug in the expunge command
Expunging from past the end of the message list (e.g., the key strokes
d > > x) no longer brings you to the first message.  Additionally, the
code in owl_function_expunge that is relevant to redisplaying after
messages are removed from the message list was placed in helper
functions, in preparation for the next commit.