Removed some NULL checks for things that really shouldn't ever be NULL
If a BarnOwl variable can't be gotten, then something is horribly wrong,
and if it can't be gotten as a string, then perl will be confused (and
something is probably horribly wrong anyway).

Additionally, our variable descriptions shoud be descriptive.  Valid
settings should not be NULL (and NULL is more likely indicative of
forgetting to set it than actually wanting an empty valid set).