Use getpwnam_r instead of getpwnam
It's re-entrant, so that's probably a good thing in case some other
thread calls getpwnam. More relevantly, Solaris has both a standard and
a non-standard version, and perl redefines getpwnam to a getpwnam_r. We
manage to convince Solaris to use the standard form, but perl doesn't
seem to notice and defines the macro wrong.

So just avoid the function altogether. Seems as reasonable of an excuse
to use the ostensibly better (and infinitely less friendly) one as any.