Use g_build_filename instead of g_strdup_printf to build paths
Saves us a duplicate slash in a few places, although most of the time it
doesn't do anything. Eh, may as well use it and be more explicit or
something.

Also change the one existing call to g_build_path to g_build_filename
instead. It's much less of a pain to use.