Refactored perl variable creation code in preparation for ambiguous types
When we allow enums in perl, we can't be sure that recreation of an int
isn't actually changing it into an enum (though, really, variables
should only be recreated when you reload module code).  This makes
BarnOwl reinitilize all the fields, but keep the current value, if there
is one.