Clamp the number of color pairs even on ncurses ABI 6
We should be able to handle color pairs past 256 on ABI 6 (which is most
distributions but Debian-based ones), but there are more changes that
need to be done. We cannot use COLOR_PAIR() and OR it into an attr_t.
This means we also can't use wbkgdset and must use wbkgrdset and
setcchar, but that does not seem to be sufficient. It's possible ncurses
is buggy here.

For now, just use the same hack as, either way, we are not ready for
ext-colors yet.