Correctly display multiline fields in :info
Commit 259e60a8abc2d2d41bc7a484b02a334fdd13857f removed the length limit
on :info, but kept the replacement of '\n' with '~' and '\r' with '!'.
This replacement came with the initial commit, with a comment of "/*
just for testing for now */".  This commit does away with that
replacement, and indents the fields for display purposes.

This is particularly useful for messages that have very long fields,
such as the original xml field for jabber messages.

Trailing newlines (e.g., in formats or in the message body) are doubled,
because the code is simpler that way, and because I don't see a
compelling reason to strip them.