Flush the fmttext cache and redisplay on setting colorztext
This fixes trac-#147.

owl_function_calculate_topmsg should probably hang off the mainwin.
Currently, the implementation relies on the existence of the current
view and recwinlines (which relies on the existence of the global
mainpanel).  The global mainpanel is guarenteed to exist if the mainwin
has been created.  The scrollmode variable must also have already been
created (it is currently safe to assume that if the mainwin exists, then
all the variables have been created).