Don't spin CPU in AIM connections
Embarrasing little bug. We were inheriting the G_IO_OUT from the login
code in the branch below and thus waking up from the event loop all the
time.

Also adjust the comment for G_IO_OUT. It makes it sound like far more of
a hack than it really is.

Reported-by: Edward Z. Yang <ezyang@mit.edu>