Workaround color-pairs and broken build on Solaris locker build
This reverts commit a71b4e049b7d1a44ddd3f918e67c45e2f8bbef2f. Really
should have verified the Solaris build first. Everything is ready for us
to not need this change, but the locker build of ncurses needs to be
updated. In the meantime, hack around the bug by clamping color pairs.

Also revert wbkgrndset because it requires cchar_t. That gets defined on
_XOPEN_SOURCE_EXTENDED or _XOPEN_SOURCE>=500. Unfortunately, the check
for the latter was added after the locker version, and Solaris is a
stickler about flags so _XOPEN_SOURCE>=500 and _XOPEN_SOURCE_EXTENDED
conflict.

Revert this change when we bundle a newer ncurses.