Allow building a zwrite from an argv
Add new versions of owl_zwrite_create and owl_zwrite_new. Note that this
means we rebuild z->zwriteline from scratch unconditionally. The only
behavior change is that the user's spacing and quoting is not preserved
in zwrite locktext.  For instance

  :zwrite     davidben       -i   "spaces yay?"

now gets a locktext of

  ----> zwrite davidben -i 'spaces yay?'

I think this isn't a big deal. It already happens for every other *write
locktext, and if we kill the buff argument, it'll happen there too.