Switch to interactive context before sourcing the startup file
There are no invariants that don't hold before the previous place we
flip the interactive bit; although default_style is applied afterwards,
there is still a style set before that code.

There is no code which reads the READCONFIG bit, so no behavior should
change from there. Setting the interactive bit makes
owl_function_load{login,}subs noisier, but that's fine. A number of
commands were labeled OWL_CTX_INTERACTIVE, but that doesn't change much
because, unless the command is OWLCMD_*_CTX, the bits are ignored.

What does change is that .owl/startup is allowed to leave with a context
on the stack. In particular, this /finally/ fixes #161.

More fundamentally, this makes .owl/startup identical to the user having
run those commands in succession, which is a nice simplification.