Unbundle random Facebook module libraries
Probably saner to ship them in the locker, lest some random other CPAN
module pick them up as dependencies. reload-module seems to get kind of
confused when there's a copy in a PAR and in the locker. The original
motivation was because they weren't in Ubuntu but

a) You can just install them from CPAN.
b) They're in precise now.
c) The locker doesn't care.

Facebook module still bundled because it's forked.