Fix the sense of drop_failed_logs
Previously, :flush-logs was dropping failed logs, while :flush-logs -f
was not.  Additionally, :flush-logs -f was enqueing the failed messages
at the end of the queue, breaking message ordering.

This commit fixes both of these issues.  I chose to optimize for the
case where :flush-logs is run without renewing tickets (where the
correct behavior is to stop at the first message that fails), rather
than the case where the user is logging to a directory they never have
permission to access (where the correct behavior is to try to log all of
the messages, and keep only the ones that fail).  This way, we won't
ever flood the system with lots of requests for files we don't have
permissions to access, and the user will have to :flush-logs -f if they
log to a directory they don't have permission to write to.